"11:20am.
```fs
        | FileChanged x ->
            // The multiple edits given here all point to coordinates in the original array, but tokenizer works sequentially.
            // So I am reversing them from last to first here.
            (let res = IVar() in Ch.give (server_blockizer x.spiPath) (Req.Modify(Array.rev x.spiEdits,res)) >>=. IVar.read res)
            |> run |> Json.serialize
```

I forgot about this. Didn't I decided to just use the range provider?

```fs
| Modify(edits,res) -> Array.iter replace edits; IVar.fill res errors
```

Ah, it just sends back tokenization errors. Right.

11:30am.

```fs
        | FileChanged x ->
            // The multiple edits given here all point to coordinates in the original array, but tokenizer works sequentially.
            // So I am reversing them from last to first here.
            (let res = IVar() in Ch.give (server_blockizer x.spiPath) (Req.Modify(Array.rev x.spiEdits,res)) >>=. IVar.read res)
            |> run |> Json.serialize
```

I had to think about this for a while. As an icebreaker, this issue will be the first on the TODO list.

The problem here is that I have no guarantee about the order VS Code is giving me for anything. So I would be better just sending all the lines as one big change.

Yes, that means on a rename where something on the start of a file and at the end of file changes it will be sending the whole file instead of just two lines, but the extra roboustness is worth it.

Let me do this.

Before I start, let me run the server again just to make sure it still works.

```ts
    const spiChange = async (doc : TextDocument, changes: readonly TextDocumentContentChangeEvent[]) => {
        const edits = changes.map(x => {
            const from = x.range.start.line
            const nearTo = from + numberOfLines(x.text)
            const lines : string [] = []
            for (let i = from; i < nearTo; i++) { lines.push(doc.lineAt(i).text) }
            return {lines, from, nearTo: x.range.end.line+1}
        })
        errorsSet(doc)(await spiChangeReq(doc.uri.fsPath, edits))
    }
```

This is what I need to work on.

11:45am.

```ts
        const from = changes.reduce((s,x) => s < x.range.start.line ? s : x.range.start.line, doc.lineCount-1)
        const nearTo = changes.reduce((s,x) => x.range.end.line < s ? s : x.range.end.line, 0)+1
        const edits = changes.map(x => {
            const from = x.range.start.line
            const nearTo = from + numberOfLines(x.text)
            const lines : string [] = []
            for (let i = from; i < nearTo; i++) { lines.push(doc.lineAt(i).text) }
            return {lines, from, nearTo: x.range.end.line+1}
        })
```

No this is not going. I just realized that it is not that easy to do this. While it is easy enough to get the removal range, what about the lines added. I can't ignore this issue either and leave it as is.

```
a b c
```

Suppose I replace `a` with `q\nw` and `c` with `r\nt`. Then I do not think the above will work correctly. The above only covers the case where there are multiple edits on diffent lines.

11:55am. I could optimize it a little if I could calculate the prefix and suffix lengths of the new array.

```fs
        const from = changes.reduce((s,x) => s < x.range.start.line ? s : x.range.start.line, doc.lineCount-1)
        const nearTo = changes.reduce((s,x) => x.range.end.line < s ? s : x.range.end.line, 0)+1
```

Though this gives me the range, I can't use this to do that because I do not know the line count of the old document. At most I could do the prefix.

12pm. I am thinking, when there is just a single edit, send that, but when there are multiple ones, send the whole file instead.

How about that?

I am not going to start keeping track of the document line counts in the client. So that seems like the most realistic solution.

12:05pm.

```ts
    const spiChange = async (doc : TextDocument, changes: readonly TextDocumentContentChangeEvent[]) => {
        if (1 < changes.length) {
            await spiOpen(doc)
        } else {
            const x = changes[0]
            const from = x.range.start.line
            const nearTo = from + numberOfLines(x.text)
            const lines : string [] = []
            for (let i = from; i < nearTo; i++) { lines.push(doc.lineAt(i).text) }
            const edit = {lines, from, nearTo: x.range.end.line+1}
            errorsSet(doc)(await spiChangeReq(doc.uri.fsPath, edit))
        }
    }
```

I should have done this the first time around. Doing these changes was pretty trivial, but now I have a correctness guarantee. Let me just run it to make sure serialization still works.

12:25pm. Doing a little Reddit posting. Bindings are a story close to my heart. Maybe in the future, I'll put in extra effort for integrating Spiral with .NET. For example, maybe I could autogenerate bindings for the .NET standard library functions and things of that nature.

12:30pm. Ok, this little bit of effort warmed me up towards the server again.

Let me have breakfast and I will seriously start thinking about how to integrate the parser with the server.

```
Spiral plugin is active.
index.js:27
rejected promise not handled within 1 second: TypeError: Cannot read property 'range' of undefined
extensionHostProcess.js:976
stack trace: TypeError: Cannot read property 'range' of undefined
```

Oh, what is this?

```fs
} else if (1 === changes.length) {
```

Let me add this check. It seems it will send empty changes to the function. How lame of VS Code to do this."