"9:50am. Let me chill for a while. I think I had my fill of games for a while. I've decided - for the next two weeks or so, my focus will be fully on programming. My dissatisfaction is on the rise, and I am going to channel that desire into programming once more.
This night, I actually nurtured my desire to do the task at hand rather than being laid out from all the demon slaying. Doom was really good for a while, but now it is time to give it a rest.

9:55am. That having said, my sleep tonight was intermittent, and I feel very groggy right now. I do not feel like starting just yet.

10:10am. Right now I am just thinking about things. Not about the thing from yesterday. I think that somehow I've managed to beat that, and hopefully I'll have internalized the need for social rules. It seems strange to do it like this, but I am going to dedicate the next two weeks of programming in their honor. I certainly won't have the opportunity to use them often, so I'll have to cement them in a different fashion.

As a kid I always had stupid rules I could never follow, but good rules can be followed. I'll be able to know that those rules are a part of myself when I am put into a situation and have done an action because of such a rule. I mean, they are really simple, not more complicated than wearing a seatbelt in the car, but it is important that I follow them from here on out. It is important that I believe in them. It is important that they be formalized ahead of time.

10:15am. At any rate, that is not really what I am thinking right now.

10:20am. I do not want to say it out loud, because it is 100% sure to give the wrong image, but I am thinking about an essay on transexualism. The subject starts there and goes in interesting directions.

I think there is a language that has not been reified yet there, in such a way of thought.

10:30am. I feel a certain sense of inspiration, and can even dig it out a little, but not completely. Those flames warm my cold and anguished heart.

Damn it, I want to agonize over this today. I want to brainstorm.

Yesterday I did not feel like programming much. I gave it a try according to my principles, but then I got hijacked. Today I definitely feel more, but I am not all there yet.

I need to be completely fed up with everything else - then I will start programming.

This next part of editor support is a major thing. It is completely an auxiliary, but will make a difference in usability of the language. It is fruit that I cannot afford not to pick. I need to make up my mind to just do it. I need to find the desire.

10:40am. If you want to accomplish something, but do not have the desire, then finding that desire is the first step you need to take care of. Cut of all distractors, embrace solitiude and look for it inwardly.

We all know the Singularity is going to happen. The hard part is acting like that is reality.

11:50am. Ok, here is the plan. I'll spend the rest of the day in bed just brainstorming.

I had to go through this in order to prime myself for the parser as well, and it seems auxiliaries won't be an exception. Saying it is simple or trite and trying to cajole myself into it that way does not work. There is a lot of work to be done and I need to steel my heart for it.

Until I manage to accomplish this, I'll stay away from the screen."