"9:15am. Got up 10m ago. What I want to do right now is just slack. I do not feel like programming. Ok, variable coloring won't be hard. But reachability highlighting will be harder.
Figuring out which variable is reachable is easy. What I need to think about is how exactly to communicate that information. What do I put the positional information in the parsed AST and so on. I'll have to grapple with this from here on out. But since I've reached a milestone of finishing the parser, first I want to slack, then I want to rev up my engine again. Screw programming every day without rest.

10:25am. I want to do more slacking. This is the ultimate luxury.

10:45am. Ok, let me start. It is time to start getting mentally ready for the next steps.

Start the IDE.

The slack feeling that I had yesterday when I finished was great. I feel I've been amply rewarded. But now it is time to start raising the tension.

10:50am. How about I step away from the screen. Like I am not lethargic in the same sense I was a few days ago where I really needed some shock treatment, but I do feel some resistance at starting a new thing.

I need some time to think about what exactly am I going to do, sure, but I also need to convince myself that I really want to do this.

10:55am. Yeah, let me move away from the bustle and look inwards for a while. I'll shut down the computer for the rest of the day. And then I will charge up my power again and let loose.

In a way what I am doing is the opposite to meditation. Whereas mediation promotes inner tranquility, my aim is to enter a state of maximum chaos.

Just this last bit of editor support and I will quite close to done with the new language. It is one of the new things coming in v0.2. Once I have this, I'll deal with the prepass and the peval without much trouble and will be able to use the language. That will leave typechecking for last.

I am close to having something usable. I just need to grasp it. First in my mind, and then in reality. The most important thing as usual is the desire.

11:10am. This will be the last great project for me as a human. I will finish Spiral v0.2, and then focus on ML from there on out. I won't be doing anymore ambitious language work after that. Sure, I will write libraries in Spiral, and might do improvements to the compiler, but by the end of 2020 and probably sooner, the language will be done.

I won't be doing any more language research if I can help it simply because it is too low reward and too high effort. As good as Spiral is, if it weren't for GPUs and now neurochips, I would not have bothered with it either. It just had too much going for it given the context for me to ignore it.

But if I could get technological enhancements, enough to be able to do 1000x as much work per time unit than I can now, that would change the risk/reward calculation substantially. In that case, language work would once again take priority.

Though 1000x would be such a huge leap in productivity that it is scarcely comprehensible. Even 2-3x would be enough to make me adjust my behavior. It is just that 1000x are realistic numbers to push out when it comes to this kind of fiction.

11:15am. Let me step away from the screen as per plan. I will do these kinds of moves more proactively from here on out. I always want to keep going until my spirit collapses and it has not been serving me well.

I've always been using this technique unconsciously - if I want to do something, but do not feel like it I would step away from the screen and brainstorm, but after this much experience I can see it as an actual technique and not a part of the flow of things. I bet I could use it in contexts other than programming should I need to."