"6:35pm. I am charging things up. As I struggle in my mind the ideas are coming to me, but I need a bit more.
At first I thought of just doing typechecking, but I am starting to change my mind again. The real target is not so much for me to typecheck a particular AST, but to flatten that AST into the language of constraints.

One thing that is giving me trouble are ranges - right none of the nodes in the parsed AST have ranges and I struggling to figure out where to put them, but if I forget about that for a moment and imagine such an AST being mapped to a bunch of equality constraints all of which have ranges, things become a lot easier.

I also gave it a lot of thought how to deal with prototype constraints. Forget dependent constraints - at this point I've completely forgotten what they even are.

6:45pm. Forget interpreting as I go along. My number one goal should be to map the parsed AST to equality constraints. After that I'll have to figure out how to do let generalization. I'll figure that out along the way as well.

On one hand just typechecking is easy enough, but I really really need to fill in the implicits automatically. No point in doing anything without that.

I am going to stop for the day here. Tomorrow I'll resume brainstorming. Right now I do not feel like raising my tension until I hit my limit. It is not really a move that I can play all the time. I'll grind through this eventually.

I'll definitely do it. Once I master type inference, the rest of the language will be a piece of cake."