"8:50am. Let me chill for a while and I will start.
9:10am. I completely lack tension right now. Let me finish my morning reading and I'll put some words in.

9:50am. Let me start. It is time to get something done.

First of all, I need to wreck the parser. It is time that I start including the ranges of everything in the AST.

Right now, I actually feel strained because I do not have that. It should be an easy decision to just do that, but internally I am bemoaning the loss of modularity that will arise from that. If there is one thing I need to say to myself that would be - get over it.

9:50am. Hmmm, before I start, I do want to do one thing. I really miss having that last statement in a block error message.

```fs
let stmts = many1 (if_ (=) statement_parsers) .>>. ((if_ (<=) (skip_keyword SpecIn) >>. root_term) <|> if_ (=) next)
```

I have to make this last part optional and then check for it.

10am.

```fs
| ExpectedExpression -> "A sequence of statements should end in an expression."
```

Calling it a block would just cause confusion so I redid the error message.

```fs
        let stmts =
            many1 (if_ (=) (range statement_parsers)) .>>. opt ((if_ (<=) (skip_keyword SpecIn) >>. root_term) <|> if_ (=) next)
            >>= fun (a,b) _ -> match b with Some b -> Ok(a,b) | None -> Error [List.last a |> fst, ExpectedExpression]
        let expr = if_ (=) next |>> fun x -> [],x
        (many1 (stmts <|> expr)
        |>> fun x ->
            List.foldBack (fun (stmts,expr) s ->
                let process_statements s = List.foldBack (fun (_,a) b -> a b) stmts s
                match s with
                | ValueNone -> ValueSome (process_statements expr)
                | ValueSome expr' -> ValueSome (process_statements (l PatE (unop ErrorNonUnit expr) expr'))
                ) x ValueNone |> ValueOption.get
            ) d
```

Let me give this a try.

Yeah wonderful, it works.

10:05am. Now...

It is time for the second phase of the parser. I need to annotate the entire AST with the node ranges. I'll also draw out Apply and PairCreate as separate nodes since the typechecker will need to recognize them. Same for the if statement and maybe some other things. Regular Ops will not have a type inferred and will instead just return a metavariable.

10:10am. I'll do this thing properly and to the best of my ability.

Before I can even start on the next phase, I need to make sure that the parser sets up the AST in such a way that I have all the information that I need to operate on it.

This is something that has been causing me a lot of consternation, because I keep wanting to have the ranges, without actually having to do the work in the parser to get them. Back when I was just doing partial evaluation I managed to get away with this, but not anymore.

10:20am. There are a lot of technical challenges when it comes to the 'typechecking' phase, but for today I will just focus on setting up the AST.

Since I am done with the parser, I might as well do this. Once I get this out of the way, I will find it easier to think about the rest. Every time I do, I get distracted by my thoughts swarming to the subject of positional information.

10:25am. Ok, let me commit here and I'll start work."