"11:30am. Yeah, I think in review you will always regret not being rational in the end. I think this is one lesson I can take from my decades of life.
12:30pm. Done with breakfast.

What I'll do today is do the monthly review. What did I even do this month?

Blockization and parsing after that. It seems I've been working on the parser for 2.5 weeks now. I has been longer than I thought it has. Yeah, progress is being made slowly. Let me start the review with this.

///

In July, I made the tokenizer incremental, finished the blockizer and then made the block parser. Rather than redoing the whole file on every change, in v0.2 the file is split into blocks based on where statements begin and end. Meaning, every top-level statement is treated as its own block independent of the rest.

In the previous version of Spiral (and the January v0.2 prototype) I did not spend too much effort on making parser errors good, in fact it really bothered me how arbitrary they appeared to be during testing, but what I have now is worlds apart. Start the server, and start the VS Code plugin and you can see the errors exactly on point in the editor now. I do not think the error messages are great, but they are not poor by any means. If you know the language syntax they are quite satisfactory. The language syntax requires almost no backtracking, so I've been able to take advantage of that to good effect.

If I were competing with F# for which language has better tooling I'd have quite a hard time. The same goes with Haskell when it comes to type system features. I've made up my mind to ignore those temptations and just focus on getting the product to ship. The main goal with v0.2 is to make something broadly useful, and then if I can get those ensemble agents to work, only then will I focus on polishing up the language. I must not forget that the killer use of type systems is that they enable top-down reasoning. Once you have a working type system and incremental compilation, you get 85% of the benefit of a language. Things like auto complete, unused variable highlighting, renaming, reference finding...plus esoteric type system features that have substitutes, all the things that while being beneficial would also take a lot of work can be left for the future.

The next thing on my list is the top-down type system. At the time of writing this, I just finished annotating every part of the parsed AST with its range and thanks to that I can finally start considering the type checking phase seriously. I thought I might do the prepass and the partial evaluation pass with which I am much more familiar before this, but I've decided in the end that I do not want redo those after I go back and do type checking. So I'll do it all in order.

Mastering type inference is one of the main challenges for me in v0.2. Everything else, while a lot of work, I already know how to do. Once I push through this part, the core feature of v0.2 will be done and I will be able to put everything into place.

My estimated time until I am done is 2-3 months. I'll need at least a month to get type checking done, and then at least a week each for the prepass and the partial evaluator which is fast, but I am quite familiar with those parts. Add codegen and testing to that and it will fill out another two. More likely, this will be an underestimate. The type checker might take longer, or work that I haven't anticipated might need doing. Back in 2017 I wanted to be done with Spiral in a few months, but it took much longer than that.

Still, by the end of the year I will have Spiral in a very usable state. I am eagerly anticipating that moment. At that time, I will be able to breathe a welcome sigh of relief. This sweltering summer is much like the work I am doing right now. I want to exchange it for the coolness of fall. To me Spiral was always a means to an end, and once it is done I am looking forward to testing my vision whether Spiral is really a necessity for catching the wave of new AI hardware that will soon arrive.

///

1:50pm. This is good. Today was the ideal time to do this review.

1:55pm. For the rest of the day I will slack. I think I will take some time off away from the screen simply to charge up the motivation to do the next step."