"10:45am. Both yesterday and this morning there were bouts of bad weather that I had to deal with by staying in bed.
Anyway, that back and forth in the PL monthly thread with /u/--comedian-- seems to be done. The sticking points to him were Julia and spiking NNs. Well, as to the later, a reasonable assumption is that they will still be computers. As long as that holds, I will be able to control them. Like, I do not think they will have restrictions just for the sake of being restricted, and internally they should be von Neumann using spikes for inter core communication.

Though, I haven't looked into what communication with 'spikes' really entails. Are they strictly binary, or something else? Well, even if they are binary, you can still group them and use that to transmit floats. In the worst case, I might have to implement a messaging library like ZeroMQ specifically for spiking NNs.

As plan B, if spiking NNs turn out to be too hard to handle is to point out that many of the devices will not be spiking NNs, and will just be accelerators with on-core memory.

11:10am. Anyway, I have plenty of potential roads to go down.

11:15am. One thing that really bothers me right now are potential compile times. Again with the prototype instances. Will I really be ok?

If Spiral becomes popular, this will be an issue. I am afraid of this. I resolved GADTs, but the current way of doing prototypes very much necessitates recompiling everything from scratch every time.

11:25am. Wait, I figured it out!

Weeks ago I remember wracking my brain over the same problem, but I ended up concludding that for every type I'd also have to pass the instances explicitly. This is a prerequisite for doing that kind of implicit analysis and it would be way too slow. This is not acceptable solution.

Right now, nominals are just ids in the typechecker.

But suppose I put ids and instances together and then hash consed them to get back a single id once again?

That would work and would be fast. That would allow me to keep join points persistent between compilation runs.

This is an elegant way of implementing a very simple idea of changing the type id when some of its nominal instances change as well.

Memoization, is there anything it can't do?

11:30am. This is the holy grail of making the partial evaluator performant even in the presence of things like typeclasses, and I just found it.

11:35am. Somewhat ironically, I could not tap into this improvement if Spiral was still a command line tool like in the v0.09 era. It is only because I am doing editor support that the language compiler is a server rather than a one-off function. It is only thanks to it being a server, that it can afford to have memory between compilation runs.

11:50am. Where did the last 15m go?

At any rate, yesterday I finalized the type system design, and just now I have everything I need to make the overall design complete from a performance standpoint.

It is fear that spurned this on. Yesterday on HN I saw a Julia thread and people were complaining that individual packages sometimes take dozen of second or more to compile. This is the one thing I want to avoid most of all in Spiral, but I could not figure out to do until now.

Now, I do not think Spiral will be anywhere as slow as Julia in overall compilation, but redoing everything from scratch every time would have been a large overhead regardless of the situation. I do want to do the most I can to make the join point information between runs persistent.

12pm. Let me take a break here. There wasn't time to do any programming in the morning anyway, and I made the ideal use of this time to resolve another design issue.

12:05pm. What I will try to do from here is hone in on my desire to do type inference for terms. Maybe I'll have to spend another day in bed to do this.

I know that I did a lot of this yesterday, but rather than motivating myself to start the term level, I struggled with the issues surrounding GADTs. And it is true that I spent a lot of time brainstorming how to do unification. I was really motivated to figure out how to do all aspects of it and now I have that, but I do not have the strong urge of wanting to do it yet.

As this morning showed, rather than being motivated to start coding, I am still thinking other things. I want to take some time to think whether I really have everything to start dealing with inl statements. If I do, there will be nothing stoping me from finishing the language.

It is much easier to make progress when you have a clear road in front of you instead of curves at every turn. Have I straightened it out enough in my mind?"