"9:30am. I am going to require foralls rather than doing it like F# does.
```
inl f (x : a) = ...
```

In this case, it makes it confusing whether `a` is a forall or a metavariable. Metavariables get generalized, but foralls don't. Sure it would be possible to do it like F#, but then I'd have to put in some special metavariable that also doubles as a forall. Screw that.

Yesterday, after I went to bed I had two other ideas.

* Make paired symbol type lowercase only.
* Op fields should be parsed as bottom up.

I've instituted the first one to make union cases more what they would be called, I just realized that the consequence of the first rule is that...

```
(a: 1) + 2
```

In a type error `a: 1`'s type would be printed as `A: i32`. Agh, that is not what I had in mind. So let me make paired symbol type what it actually is.

Regarding parsing the Op fields, I should have done this from the start, but simply forgot. I'd have realized my error once I try parsing a type variable and got a parse error.

```fs
            | (r,x) :: k', v ->
                if Char.IsLower(x,0) then
                    List.reduceBack (fun a b -> RawTPair(range_of_texpr a +. range_of_texpr b,a,b))
                        (RawTSymbol(r,symbol_paired_concat ((r,x) :: k')) :: v) |> Ok
                else
                    Error [r, SymbolPairedShouldStartWithLowercase]
```

Let me go with this.

```fs
| SymbolPairedShouldStartWithLowercaseInTypeScope -> "Paired symbol should start with lowercase in type scope."
```

Actually, let me name it like this.

9:45am. Now let me go for the other point. In top-down foralls, it might be good to support kind metavars and wlidcards. Instead of assuming that `forall a.` is the same as `forall (a : *).`, it might be better to give it a kind metavar since I know how to do that now.

Well, forget that for the time being. It is not too important, and I can always do it later. I'll save it for when it is time to do constraints.

Let me take care of the Op fields.

```fs
match a with
| "!!!!" ->
    (range (read_big_var .>>. (rounds (sepBy1 (fun d -> expressions {d with is_top_down=false}) (skip_op ","))))
```

There.

9:55am. Ok, let me try this out.

```
union option x =
    | some: x
    | None

inl f x =
    inl (+) (a : i32) (b : i32) = !!!!Add(a, `b) : i32
    1+2+3+X
```

Prasing type vars in op fields works now. It is awkward how `some:` is lowercase and `None` is upper, but it can't be helped. Nevermind this little thing.

This way of doing things is right, no doubt about it.

10:05am. Ok, I know that by this point am really vacilating on starting work on the term inference. Still if what I have in mind is right it won't be too much harder than anything else I've done. Given what an integral part of v0.2 it is expected to be you'd think it would take huge amounts of code, but I expect it go quite smoothly. It won't be particularly more complex than the prepass for example. And if I consider just the unification part, rather than filling out the implicits, it should be quite simple indeed, only somewhat more than validating bound variables.

10:20am. Let me do my morning reading and then I will get on with it. I missed the Satanophany raw yesterday."