"9:40am. I am up.
9:45am. Let me chill for a bit.

Before I start, let me talk a bit.

9:50am. I've been thinking about record application. One very important thing I want in Spiral v0.2 are tensors with named dimensions at the top level.

So just like with v0.09, if I have a `tensor {a : i32; b : i32} f32`, I want to be able to write `x @ .a` and get that dimension.

It is true that I thought about this a lot, and concluded that I would need dependent constraints.

```
forall a x b c [record_apply a x c]. tensor (a * b) el -> x -> tensor (c * b) el
```

This is really convoluted type. Instead I will just write...

```
forall a x b. tensor (a * b) el -> x -> tensor (a @ x * b) el
```

Rather than a constraint, I am going to add `TyRecordApply` as a type to the language. The way unification for it will work is that it can only be unified with itself. Anything else would be a type error.

This is while it is unresolved. If both the left and the right side are known then `term_subst` will reduce the record application at the type level to its actual type and it will be able to be unified normally.

Furthermore, I'll reuse the same pass as for validating unresolved metavariables in order to ensure that there are no unreduced record applications not present in the body of the statement.

This will give me what I want.

10am. I really want to not have to drop into the bottom up segment just to make use of tensors with named dimensions.

Just going a little step further will make the language a lot more suitable for numerical applications. Extended tensors are really useful - they are the reason why I could make the fully optimized RNN in an elegant fashion.

10:10am. Dependent constraints are out - I can't implement them as the type inference for Spiral is flow sensitive. Constraints that just do some validation at the end are fine, but complex constraints the operate on multiple inputs or return a type like the record one would need to be are out. I can't deal with them.

Spiral's type system needs to be simple. And I will have what I want in that regard.

10:15am. Now let me chill. After that I'll finish the last pattern case.

10:25am. Ok, enough. Since I am already out my earlier urge to slack, let me to implement that last case. After that I'll have breakfast and do some proper laziness.

Regarding that record apply, I think it will be just fine to wait until I am done with the language and am getting to the constraints. That is the best way. I do not really have to concern myself with this now as the rules are straightforward. I'll forget them until the time is ripe.

Now, let me start getting to the end of the `infer` function.

```fs
//| PatNominal of Range * (Range * VarString) * Pattern
```

Let me do this thing.

10:35am.

```fs
            | PatNominal(_,(r,a),b) ->
                match Map.tryFind a top_env.ty with
                | Some x ->
                    match head_hoc x with
                    | ValueSome i -> failwith "ok"
                    | ValueNone -> errors.Add(r,TypeInGlobalEnvIsNotNominal x); f (fresh_var()) b
                | _ -> errors.Add(r,UnboundVariable); f (fresh_var()) b
```

This time I went with the error cases.

10:45am.

```fs
            | PatNominal(_,(r,name),a) ->
                match Map.tryFind name top_env.ty with
                | Some x ->
                    match ho_index x with
                    | ValueSome i ->
                        match aux.hoc.[i] with
                        | HOCNominal(vars,v) -> let x,m = ho_make i vars in unify r s x; f (subst m v) a
                        | HOCUnion _ -> errors.Add(r,UnionInPatternNominal i); f (fresh_var()) a
                    | ValueNone -> errors.Add(r,TypeInGlobalEnvIsNotNominal x); f (fresh_var()) a
                | _ -> errors.Add(r,UnboundVariable); f (fresh_var()) a
```

This should do it. I did some renaming of `head` and `make_higher_order`.

10:50am. I am just taking a little breather. This really went quite smoothly.

```fs
        //| RawMatch of Range * body: RawExpr * (Pattern * RawExpr) list
        //| RawRecBlock of Range * ((Range * VarString) * RawExpr) list * on_succ: RawExpr // The bodies of a block must be RawInl or RawForall.
```

I am done with patterns. Only these two are left now. Let me chill and I will start work on this. It might take me a while to get through these two. Let generalization won't be hard on its own, but I will need to do validation afterwards and other checks besides.

Maybe it will take me longer than today to go through it all. But once I am done, I really will be mostly done with the typechecker. Whatever will remain should be easy to deal with.

Let me have breakfast."