"9:20am. I went to bed very late so I am very groggy right now.
The recent language design issues leave me stressed. Actually, I need constraints in order to deal with record apply properly, just not dependent ones.

The fact that need to iterate in order for all contraints to be propagated is making my brain churn. Can I do something efficient to gather all the constraints in one go? I have no idea.

9:30am. It is making me stressed. There is no easy way to validate record apply. Unless it is obviously wrong, I'll have to let them pass during metavariable validation.

But leaving out the constraints, I am quite close to being done.

9:35am. I have to admit, one of the reasons I am going to bed so late is because I am hooked on a shitty Chinese webnovel. Devil's Evolution Catalogue's premise where the MC reincarnates as a devil in hell is just too much up my alley. The shitty village arc has really been testing my patience though.

9:40am. I think I really will just leave constraints out of the language. The partial evaluator can take care of it. I just have so many other more important things to think aobut.

...Ok, as the first task of the day, let me change the parser for term foralls. I am going to add kind wildcard to the mix.

Before I start let chill.

10:25am. Let me finally finish this arc, just two more chapters and then I will do some programming.

10:45am. Ok, enough. Let me do some programming.

```fs
type RawKindExpr =
    | RawKindStar
    | RawKindFun of RawKindExpr * RawKindExpr
```

First of all, let me add the `RawKindWildcard` here.

```fs
type RawKindExpr =
    | RawKindWildcard
```

I'll just assume these are regular one stars in type definitions, but in term foralls, I will have them be kind metavars.

```
inl f forall (x : * -> *)
```

I do not want to force kind annotations everywhere.

```fs
let rec typevar = function
    | RawKindWildcard | RawKindStar -> KindStar
    | RawKindFun(a,b) -> KindFun(typevar a, typevar b)
```

This is actually the only green spot.

I will have to make an alternative for `RawForall`.

One thing that is stressing me out is how will I deal with kind in the partial evaluator. Agh...I'll deal with that when it comes to it.

```fs
        | RawForall(r,(_,(name,k)),b) ->
            let rec typevar = function
                | RawKindWildcard -> fresh_kind()
                | RawKindStar -> KindStar
                | RawKindFun(a,b) -> KindFun(typevar a, typevar b)
            let k = typevar k
            let i,v = fresh_var'' k
            forall_scopes.Add(name,i)
            let body = fresh_var()
            unify r s (TyForall((name,k),body))
            term {env with ty = Map.add name v env.ty} body b
```

Let me go with this.

Now even though I have this here, I will have to pull this out and put it into the let statements.

11am. Hmmmmm...

I have an idea.

I've been obsessed with doing the maximum amount of let generalization possible.

```
inl f x =
    inl y = ...
    inl g z = y + z
    ...
    g y x
```

...There might be an alternative to doing no local let generalization.

Right now I am obsessed with generalizing as soon as `g` has finished processing.

But suppose I generalize at the end?

Suppose `y` is still a metavariable at the end?

...No it would not work.

The only way I can unambiguously let generalize at the end is when all the top level metavar substitutions are done, but a function still has uninferred arguments.

11:10am. You know what, let me ask about let generalization on the PL sub. Maybe somebody will have some insight there.

```fs
let f x =
    let y = failwith ""
    let g z = y + z
    g (y : int) x
```

I'll also ask about the constraints.

11:15am.

///

```fs
let f x =
    let y = failwith ""
    let g z = y + z
    g (y : int) x
```

I though let generalization would be easy, but it turns out that substituting the leftover metavariables for foralls in the statement body only makes sense if the metavariables do not have paths connecting them to other metavariables outside their scope.

Here

///

...This is not a good example.

Of course if `y` gets inferred as `int`, then `z` should be `int` as well.

Let me try this instead.

```fs
let f x =
    let g y = y
    g 1
```

11:20am. Great, now I can figure out how to pose the question.

```fs
let f x =
    let y = failwith ""
    let g z = y + z
    g (y : int) x
```

11:55am.

///

I though let generalization would be easy, but it turns out that substituting the leftover metavariables for foralls in the statement body only makes sense if the metavariables do not have paths connecting them to other metavariables outside their scope.

```fs
let f x =
    let y = failwith ""
    let g z = y + z
    g (y : int)
```

Here, when type inferencer finishes with `g` it would have a type like `'a -> 'a`. But I cannot generalize this to `forall a. a -> a` because `y` might turn out to have a type like `int` down the road. So the correct thing to do this is to just let the inference proceed, and `g` will have the `int -> int` type at the end.

```fs
let f x =
    let g z = z + z
    g (x : int)
```

Here, `g` would get inferred as `'a -> 'a` once again, but since `'a` does not have any paths outside its scope, I should be able to generalize it as a function of `forall a. a -> a`.

The issue for me here is that I do not want go all over the metavariable array doing an in depth search if any metavariable outside the scope of `g` connects to it. It does not really feel like the right choice of action here. It is just seems too inexpensive and inelegant as a solution.

In my own language it is possible to simply do...

```fs
let f x =
    let g forall a. (z : a) = z + z
    g (x : int)
```

...and this would force it to generalize. This kind of insight is not what I am looking for though. What I want is to automate cheaply as much as possible.

Is there any way to resolve the inherent tension between wanting to generalize as soon as possible and leaving it for after the inference is done? Are there any insights that can be used to make this easier?

///

Let me go with this.

12pm. https://www.reddit.com/r/ProgrammingLanguages/comments/i71zyx/how_to_do_generalization_of_local_let_statements/

I posted it on the PL sub. But I do not have much hope on getting any good replies.

The idea just now of leaving let generalization for the last is a good one though. This would be pretty decent actually. It is good enough.

12:10pm. I am thinking.

Even though just started an hour ago, let me have breakfast here. No point in starting anything else right now.

All this really is stressful. From a distance something like let generalization seems easier than it is. Once you get closer it is only then that you see all the details that need taking care of and the lethargy hits you.

Well, programming is usually like that. Forget eager let generalization and constraints. I'll do the deferred version, and who knows, maybe for both constraints and generalization, an idea will hit me down the road, much like it did for the prepass.

I really do not want to make this decisions, but for v0.2 I won't be doing constraints in order to lower my workload. But on the positive side, that means that once I am done with what I need to do now, I will be pretty much done with the language.

I do not want to have shitload of stuff that need doing waiting for me after I am done with the typechecker here. I want to be finished with v0.2. The TC can act as support and the partial evaluator can do the rest of the heavy lifting.

Let me put in `RecordApply` at the type level in now. It is easy enough and I do not want to wait until after let generalization to do it.

I do not want to backtrack. After I am done with the last two cases, I want to start finishing up the typechecking phase and move to testing.

I'll do that after breakfast."